internal_services/requirements.txt
wheel
untangle
sympy
xmlschema==1.0.14
django
deepdiff
defusedxml
ipython
django-modern-rpc
